refactor: directory changes
