+++ src/main/kotlin/ru/quipy/projections/ProjectEventsSubscriber.kt
@@ -4,10 +4,9 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
-import ru.quipy.api.ProjectAggregate
-import ru.quipy.api.TagAssignedToTaskEvent
-import ru.quipy.api.TagCreatedEvent
-import ru.quipy.api.TaskCreatedEvent
+import ru.quipy.api.aggregates.TaskAggregate
+import ru.quipy.api.aggregatesEvents.TaskCreatedEvent
+import ru.quipy.api.aggregatesEvents.TaskStatusChangedEvent
 import ru.quipy.streams.AggregateSubscriptionsManager
 import javax.annotation.PostConstruct
 
@@ -21,18 +20,14 @@ class ProjectEventsSubscriber {
 
     @PostConstruct
     fun init() {
-        subscriptionsManager.createSubscriber(ProjectAggregate::class, "some-meaningful-name") {
+        subscriptionsManager.createSubscriber(TaskAggregate::class, "some-meaningful-name") {
 
             `when`(TaskCreatedEvent::class) { event ->
                 logger.info("Task created: {}", event.taskName)
             }
 
-            `when`(TagCreatedEvent::class) { event ->
-                logger.info("Tag created: {}", event.tagName)
-            }
-
-            `when`(TagAssignedToTaskEvent::class) { event ->
-                logger.info("Tag {} assigned to task {}: ", event.tagId, event.taskId)
+            `when`(TaskStatusChangedEvent::class) { event ->
+                logger.info("Tag {} assigned to task {}: ", event.taskId, event.taskId)
             }
         }
     }
